prepare for branch matrix-analysis

+++ src/main/scala/edu/nju/pasalab/marlin/matrix/SparseVecMatrix.scala
@@ -23,25 +23,25 @@ class SparseVecMatrix(
   final def multiplySparse(other: SparseVecMatrix, parallelism: Int): CoordinateMatrix = {
     val partitioner = new HashPartitioner(parallelism)
     val thisEmits = rows.flatMap( t => {
-      val arr = new ArrayBuffer[(Long, (Long, Double))]
+      val arr = new ArrayBuffer[(Long, (Long, Float))]
       val len = t._2.indices.size
       for (i <- 0 until len){
-        arr += ((t._2.indices(i), (t._1, t._2.values(i))))
+        arr += ((t._2.indices(i), (t._1, t._2.values(i).toFloat)))
       }
       arr
     }).groupByKey(partitioner).cache()
 
     val otherEmits = other.rows.flatMap(t => {
-      val arr = new ArrayBuffer[(Long, (Long, Double))]()
+      val arr = new ArrayBuffer[(Long, (Long, Float))]()
       val len = t._2.indices.size
       for ( i <- 0 until len){
-        arr += ((t._1, (t._2.indices(i), t._2.values(i))))
+        arr += ((t._1, (t._2.indices(i), t._2.values(i).toFloat)))
       }
       arr
     })
 
     val result = thisEmits.join(otherEmits).flatMap( t =>{
-      val arr = new ArrayBuffer[((Long, Long),Double)]()
+      val arr = new ArrayBuffer[((Long, Long), Float)]()
       val list = t._2._1.toArray
       for (l <- list){
         arr += (((l._1, t._2._2._1), l._2 * t._2._2._2))
